// src/app/site/interfaces/imeta-data.ts
/**
 * Website - IMetaData - Interface
 *
 * @since       2018.04.22
 * @version     1.0.0.0
 * @author		  mlbors
 * @copyright	-
 */

/********************************/
/********** IMETA DATA **********/
/********************************/

export interface IMetaData {
  id: string;
  name: string;
  content: any;
}